added a late counter, timeout command, and changed the yikes command

+++ cogs/kill/off.py
@@ -3,14 +3,35 @@ from discord.ext import commands
 
 
 class Kill(commands.Cog):
-    """Turns the bot off"""
+    """Commands For Caleb's Use"""
 
     def __init__(self, client):
         self.client = client
 
+    
+    @commands.command()
+    async def timeout(self, ctx, *max_time):
+        """Deletes all messages for a given period of time
+        can pass in the amount of time as a variable `$timeout <time>`"""
+        import time
+
+        if not max_time:
+            max_time = 600
+        else:
+            max_time = max_time[0]
+
+        await ctx.channel.send(f'**Due to conflict:** Chat is disabled for {max_time} seconds')
+
+        start_time = time.time()  # remember when we started
+        while (time.time() - start_time) < int(max_time):
+            thing = await self.client.wait_for('message')
+            await thing.delete()
+
+        await ctx.channel.send(f'Chat is enabled. Please be friends.')
+    
     @commands.command()
     async def off(self, ctx):
-        print(ctx.message.author.name)
+        """Turns the Bot off"""
         if ctx.message.author.name == 'CalebP':
             await ctx.channel.send(f'https://tenor.com/view/cry-sad-toy-story-woody-so-long-partner-gif-9797730')
             exit()
@@ -19,10 +40,14 @@ class Kill(commands.Cog):
 
     @commands.command()
     async def restart(self, ctx):
-        await self.client.change_presence(activity=discord.Activity(type = discord.ActivityType.streaming, name = f'Rebooting...') )
-        import os
-        os.system("sudo reboot")
-        exit()
+        """Restarts the Bot"""
+        if ctx.message.author.name == 'CalebP':
+            await self.client.change_presence(activity=discord.Activity(type = discord.ActivityType.streaming, name = f'Rebooting...') )
+            import os
+            os.system("sudo reboot")
+            exit()
+        else:
+            await ctx.channel.send(f'Shut the fuck up {ctx.message.author.name}, you dont have the authority for that.')
